% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R
\name{mvnpdf}
\alias{mvnpdf}
\title{mvnpdf
this function calculate the (log)normal distribution for a matrix}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{the matrix of the observations}

\item{mean}{a vector, the mean for each variable}

\item{varcovM}{the matrix of variance, covariance}

\item{Log}{logical, default equal TRUE if TRUE calculate the log normal distribution}
}
\value{
the observation and the prediction with the (log)normal distribution
}
\description{
mvnpdf
this function calculate the (log)normal distribution for a matrix
}
\examples{
mvnpdf(x=matrix(1.96), Log=FALSE)
}
